Service refactoring. Change double to BigDecimal

+++ src/main/java/ru/payment/calc/payment_calculator/config/props/DebtThreshold.java
@@ -4,11 +4,13 @@ import lombok.Data;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.context.annotation.Configuration;
 
+import java.math.BigDecimal;
+
 @Configuration
 @ConfigurationProperties(prefix = "debt")
 @Data
 public class DebtThreshold {
 
-    private double threshold;
+    private BigDecimal threshold;
 
 }
